requirements.txt
cycler==0.10.0
matplotlib==1.5.3
mpld3==0.2
mpmath==0.19
numpy==1.11.2
PyDSTool==0.90.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-libsbml==5.13.0
pytz==2016.7
scipy==0.18.1
six==1.10.0
sympy==1.0
python-libsedml==0.4.1
